chore: add an example test suite and an example function to render
The describe defines a test suite where the first name is typically the name of the component or the function we are going to test. The test case described by the keyword 'it' perform certain checks with the help of assertive functions. It renders any valid JSX using the render function and also displays the HTML using the screen.debug()